Move UptimeGraph to shared modules for shadow+Ele
