+++ src/parser/shared/modules/getUptimeGraph.tsx
